# imaxx-smart-office-be-dev/docs/swagger.yaml
basePath: /api
definitions:
...
  request.CreateWorkflowDetailRequest:
    properties:
      is_final:
        maximum: 1
        minimum: 0
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
      role_id:
        type: string
      seq:
        type: integer
      status:
        maximum: 1
        minimum: 0
        type: integer
    required:
    - is_final
    - name
    - role_id
    - seq
    - status
    type: object
  request.CreateWorkflowRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
      status:
        maximum: 1
        minimum: 0
        type: integer
      workflow_details:
        items:
          $ref: '#/definitions/request.CreateWorkflowDetailRequest'
        type: array
    required:
    - name
    - status
    - workflow_details
    type: object
...
  request.UpdateWorkflowDetailRequest:
    properties:
      is_final:
        maximum: 1
        minimum: 0
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
      role_id:
        type: string
      seq:
        type: integer
      status:
        maximum: 1
        minimum: 0
        type: integer
    required:
    - is_final
    - name
    - role_id
    - seq
    - status
    type: object
  request.UpdateWorkflowRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
      status:
        maximum: 1
        minimum: 0
        type: integer
      workflow_details:
        items:
          $ref: '#/definitions/request.UpdateWorkflowDetailRequest'
        type: array
    required:
    - name
    - status
    - workflow_details
    type: object
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: iMaxx Office API with Swagger
  title: iMaxx Office API
  version: "1.0"
paths:
...
  /workflows:
    get:
      consumes:
      - application/json
      description: find all workflow
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindAllWorkflow workflow
      tags:
      - workflows
    post:
      consumes:
      - application/json
      description: create workflow
      parameters:
      - description: create workflow payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateWorkflowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: CreateWorkflow workflow
      tags:
      - workflows
  /workflows/{id}:
    delete:
      consumes:
      - application/json
      description: delete workflow
      parameters:
      - description: Workflow ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: DeleteWorkflow workflow
      tags:
      - workflows
    get:
      consumes:
      - application/json
      description: find by id workflow
      parameters:
      - description: Workflow ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindById workflow
      tags:
      - workflows
    put:
      consumes:
      - application/json
      description: update workflow
      parameters:
      - description: update workflow payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateWorkflowRequest'
      - description: Workflow ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: UpdateWorkflow workflow
      tags:
      - workflows
  /workflows/active:
    get:
      consumes:
      - application/json
      description: find all active workflow
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindActiveAllWorkflow workflow
      tags:
      - workflows
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"